rgw: move mdlog trimming into rgw_trim_mdlog.cc
